🚚 Rename: entity로 변경
