+++ src/main/java/com/example/egobook_be/domain/letters/PlazaLetterDummyDataLoader.java
@@ -1,8 +1,8 @@
 package com.example.egobook_be.domain.letters;
 
-import com.example.egobook_be.domain.letters.domain.PlazaLetter;
-import com.example.egobook_be.domain.letters.domain.PlazaLetterMode;
-import com.example.egobook_be.domain.letters.domain.PlazaLetterStatus;
+import com.example.egobook_be.domain.letters.entity.PlazaLetter;
+import com.example.egobook_be.domain.letters.entity.PlazaLetterMode;
+import com.example.egobook_be.domain.letters.entity.PlazaLetterStatus;
 import com.example.egobook_be.domain.letters.repository.PlazaLetterRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
