+++ src/main/java/com/example/egobook_be/domain/letters/repository/PlazaLetterRepository.java
@@ -1,7 +1,7 @@
 package com.example.egobook_be.domain.letters.repository;
 
-import com.example.egobook_be.domain.letters.domain.PlazaLetter;
-import com.example.egobook_be.domain.letters.domain.PlazaLetterStatus;
+import com.example.egobook_be.domain.letters.entity.PlazaLetter;
+import com.example.egobook_be.domain.letters.entity.PlazaLetterStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -9,7 +9,6 @@ import org.springframework.data.jpa.repository.*;
 import org.springframework.data.repository.query.Param;
 
 import java.time.OffsetDateTime;
-import java.util.List;
 import java.util.Optional;
 
 public interface PlazaLetterRepository extends JpaRepository<PlazaLetter, Long> {
