+++ src/main/java/com/example/egobook_be/domain/letters/repository/PlazaLetterThreadRepository.java
@@ -1,6 +1,6 @@
 package com.example.egobook_be.domain.letters.repository;
 
-import com.example.egobook_be.domain.letters.domain.PlazaLetterThread;
+import com.example.egobook_be.domain.letters.entity.PlazaLetterThread;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlazaLetterThreadRepository extends JpaRepository<PlazaLetterThread, Long> {
